Report Metamask Wallet Error over EventBus
