+++ ui/core/src/api/EventBusService/EventBusService.ts
@@ -1,72 +1,6 @@
 import { TransactionStatus } from "../../entities";
 import { EventEmitter2 } from "eventemitter2";
-
-// Add more wallet types here as they come up
-type WalletType = "sif" | "eth";
-
-type ErrorEvent = {
-  type: "ErrorEvent";
-  payload: {
-    message: string;
-    detail?: {
-      type: "etherscan" | "info";
-      message: string;
-    };
-  };
-};
-
-type TransactionErrorEvent = {
-  type: "TransactionErrorEvent";
-  payload: {
-    txStatus: TransactionStatus;
-    message: string;
-  };
-};
-
-type WalletConnectedEvent = {
-  type: "WalletConnectedEvent";
-  payload: { walletType: WalletType; address: string };
-};
-
-type WalletDisconnectedEvent = {
-  type: "WalletDisconnectedEvent";
-  payload: { walletType: WalletType; address: string };
-};
-
-type PegTransactionPendingEvent = {
-  type: "PegTransactionPendingEvent";
-  payload: { hash: string };
-};
-
-type PegTransactionCompletedEvent = {
-  type: "PegTransactionCompletedEvent";
-  payload: {
-    hash: string;
-  };
-};
-
-type PegTransactionErrorEvent = {
-  type: "PegTransactionErrorEvent";
-  payload: {
-    txStatus: TransactionStatus;
-    message: string;
-  };
-};
-
-type NoLiquidityPoolsFoundEvent = {
-  type: "NoLiquidityPoolsFoundEvent";
-  payload: {};
-};
-
-export type AppEvent =
-  | ErrorEvent
-  | WalletConnectedEvent
-  | WalletDisconnectedEvent
-  | PegTransactionPendingEvent
-  | PegTransactionCompletedEvent
-  | NoLiquidityPoolsFoundEvent
-  | TransactionErrorEvent
-  | PegTransactionErrorEvent;
+import { AppEvent } from "./Events";
 
 export type EventHandler = (event: AppEvent) => void;
 export type AppEventType = AppEvent["type"];
