+++ ui/core/src/api/EventBusService/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./EventBusService";
 export * from "./EventBusService";
+export * from "./Events";
